#8.14 Bugfix
기존 코드에서 발생 가능했던 일종의 버그 수정

1. 비디오 업로드시 바뀐 videos배열을 푸시 후 바뀐 내용을 save할 때 미들웨어에 의해 유저의 비밀번호가 해싱(암호화)되는 현상 수정

2. 로그인만 되어있는 상태면 동영상 소유자 여부에 상관없이 영상 수정, 삭제가 가능한 현상 수정

-----------------------------------------
!!! 추가 - deleteVideo(동영상 삭제)시 Video DB 뿐 아니라 User DB에 있는 videos 배열안의 데이터도 삭제하도록 변경 (강의 댓글 참고)

+++ src/controllers/videoController.js
@@ -28,9 +28,15 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit-video", {
     pageTitle: `Edit : ${video.title}`,
@@ -39,10 +45,16 @@ export const getEdit = async (req, res) => {
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -84,7 +96,21 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  // User DB속 videos배열에 있는 정보까지 삭제
+  user.save();
   return res.redirect("/");
 };
